Jetpack: Fix OpenTable block input jumping around (#34654)
The restaurant input in the OpenTable block would change its width as
the search results changed, causing the adjacent Embed button to also
move around. Something in Gutenberg 17.2.1 changed which results show up
by default in the E2E test, causing this jumping to start making the
test fail.

The fix we're applying here is to set the input field with flex-grow and
the container with nowrap, so the field remains at the full width with
the button beside it in all cases (above the `@media` cutoff that
switches away from flexbox completely).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7214351998