Contact Form dashboard: ensure the inbox is displayed properly for all users (#35571)
* Forms: update permission check, do not check for blog membership

It is enough to check for admin permissions on the site. This will work for both single site and multisite environments.

The check removed in this commit was checking a local user ID against a WordPress.com blog ID, so this couldn't work.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7849163843